Fix for issue 57: Add methods and factory method for GGeoXml (GeoXmlOverlay)
This patch removes the constructor to GeoXmlOverlay but adds a load() factory
method.  The new factory method takes a callback which simplifies use of this
class.
  - I reorganized the JSIO implementation to use a new subclass of OverlayImpl for GeoXmlOverlay instead of using a constructor  inside the OverlayImpl class.  I don't remember seeing this pattern before, but JSIO seems to handle it just fine. 
I also added a test driver for the new methods in the KML Overlay demo in HelloMaps.

This change leaves exposed a potential race condition if the callback returns 
before the GGeoXml constructor returns an instance.  (Logged as issue 88)

Found by: strmer15
Patch by: zundel
Review by: mmendez



git-svn-id: http://gwt-google-apis.googlecode.com/svn/trunk@141 dc8c9c99-1131-0410-92e2-872299cb2494