Fixes issue 63 - Problem removing listeners from MapWidget
This patch adds a parameterized class called ListenerCollection 
that associates  an instance of a listener with the JavaScript 
event handles returned from registering events with the Maps API.  
This is used to correctly implement removeListeners().  A demo 
is provided to exercise all the listener types supported by the 
Marker class.

Patch by: zundel
Review by: mmendez



git-svn-id: http://gwt-google-apis.googlecode.com/svn/trunk@154 dc8c9c99-1131-0410-92e2-872299cb2494