fixes issue 90
Issue 90 called for refactoring the Projection class, but in the end, I determined that
the class structure was done correctly the first time.  So no changes
were needed to the API, but I did take the opportunity to create a
unit test to exercise subclassing Projection.

Patch by: zundel
Review by: mmendez



git-svn-id: http://gwt-google-apis.googlecode.com/svn/trunk@343 dc8c9c99-1131-0410-92e2-872299cb2494