fixes issue 82.
This adds the Polygon.fromEncoded() wrappers.

Of note is the construct that uses GWT.isScript() to avoid copying
the array in web mode.  Scott said this is more efficient,and
potentially there could be many polylines in a polygon.  Looking
at the generated code shows that all the wrappers for fromEncoded()
just melt away.

Patch by: zundel
Review by: mmendez



git-svn-id: http://gwt-google-apis.googlecode.com/svn/trunk@352 dc8c9c99-1131-0410-92e2-872299cb2494