fixes issue 88 - A potential race condition in the GGeoXml wrapper.
The GGeoXml constructor is asynchronous.  It is not safe to use many
of its methods until the URL data is loaded.  You can detect this via
a method call or by waiting for a callback to return.  We opted for
the second method, but had a potential race condition because the
GeoXml object might not be returned from the constructor before the
callback fires (this can be a problem on IE)

Patch by: zundel
Review by: mmendez



git-svn-id: http://gwt-google-apis.googlecode.com/svn/trunk@353 dc8c9c99-1131-0410-92e2-872299cb2494