Addresses issue 135 (update features of the Polygon object)
The most significant new features from the Maps folks are the ability 
to allow end users to edit and draw polygons.  This is just like the 
Polyline feature that was recently added (issue 86).  The change 
includes a unit test for some features and an update to the Drawing Overlay demo.

Patch by: zundel
Review by: mmendez



git-svn-id: http://gwt-google-apis.googlecode.com/svn/trunk@401 dc8c9c99-1131-0410-92e2-872299cb2494